+++ src/brasil/gov/vcge/tests/test_viewlet.py
@@ -8,7 +8,7 @@ from plone.app.testing import TEST_USER_ID
 from plone.app.testing import TEST_USER_NAME
 from zope.site.hooks import setSite
 
-import unittest2 as unittest
+import unittest
 
 
 class TestViewlet(unittest.TestCase):